+++ GRDB/FTS/FTS5Tokenizer.swift
@@ -1,4 +1,5 @@
 #if SQLITE_ENABLE_FTS5
+import Foundation
 #if SWIFT_PACKAGE
 import CSQLite
 #elseif GRDBCIPHER