+++ GRDB/QueryInterface/FTS3+QueryInterface.swift
@@ -1,4 +1,4 @@
-extension QueryInterfaceRequest {
+extension TableRequest where Self: FilteredRequest {
     
     // MARK: Full Text Search
     
@@ -11,7 +11,7 @@ extension QueryInterfaceRequest {
     ///
     /// If the search pattern is nil, the request does not match any
     /// database row.
-    public func matching(_ pattern: FTS3Pattern?) -> QueryInterfaceRequest<T> {
+    public func matching(_ pattern: FTS3Pattern?) -> Self {
         guard let pattern = pattern else {
             return none()
         }