+++ CHANGELOG.md
@@ -58,11 +58,49 @@ GRDB adheres to [Semantic Versioning](https://semver.org/), with one expection: 
 - [#562](https://github.com/groue/GRDB.swift/pull/562): Fix crash when using more than one DatabaseCollation
 - [#563](https://github.com/groue/GRDB.swift/pull/563): More tests for eager loading of hasMany associations
 - [#574](https://github.com/groue/GRDB.swift/pull/574): SwiftLint
+- [#576](https://github.com/groue/GRDB.swift/pull/576): Expose table name and full-text filtering methods to DerivableRequest
 - [#577](https://github.com/groue/GRDB.swift/pull/577): Rename `aliased(_:)` methods to `forKey(_:)`
 
 ### API Diff
 
 ```diff
+-protocol DerivableRequest: FilteredRequest, JoinableRequest, OrderedRequest, SelectionRequest { }
++protocol DerivableRequest: FilteredRequest, JoinableRequest, OrderedRequest, SelectionRequest, TableRequest { }
+
+-extension QueryInterfaceRequest {
+-    func matching(_ pattern: FTS3Pattern?) -> QueryInterfaceRequest<T>
+-}
+-extension QueryInterfaceRequest {
+-    func matching(_ pattern: FTS5Pattern?) -> QueryInterfaceRequest<T>
+-}
++extension TableRequest where Self: FilteredRequest {
++    func matching(_ pattern: FTS3Pattern?) -> Self
++}
++extension TableRequest where Self: FilteredRequest {
++    func matching(_ pattern: FTS5Pattern?) -> Self
++}
+ 
+ protocol Association: DerivableRequest {
+-    associatedtype OriginRowDecoder
++    associatedtype OriginRowDecoder: TableRecord
+ }
+
+-struct BelongsToAssociation<Origin, Destination>: AssociationToOne { }
+-struct HasManyAssociation<Origin, Destination>: AssociationToMany { }
+-struct HasManyThroughAssociation<Origin, Destination>: AssociationToMany { }
+-struct HasOneAssociation<Origin, Destination>: AssociationToOne { }
+-struct HasOneThroughAssociation<Origin, Destination>: AssociationToOne { }
++struct BelongsToAssociation<Origin: TableRecord, Destination: TableRecord>: AssociationToOne { }
++struct HasManyAssociation<Origin: TableRecord, Destination: TableRecord>: AssociationToMany { }
++struct HasManyThroughAssociation<Origin: TableRecord, Destination: TableRecord>: AssociationToMany { }
++struct HasOneAssociation<Origin: TableRecord, Destination: TableRecord>: AssociationToOne { }
++struct HasOneThroughAssociation<Origin: TableRecord, Destination: TableRecord>: AssociationToOne { }
+
+-struct QueryInterfaceRequest<T>: DerivableRequest { }
++struct QueryInterfaceRequest<T>: FilteredRequest, OrderedRequest, SelectionRequest, TableRequuest { }
++extension QueryInterfaceRequest: JoinableRequest where T: TableRecord { }
++extension QueryInterfaceRequest: DerivableRequest where T: TableRecord { }
+
  extension AssociationAggregate {
 +    @available(*, deprecated, renamed: "forKey(_:)")
      func aliased(_ name: String) -> AssociationAggregate<RowDecoder>